telegram-manager/requirements.txt
telethon
python-dotenv
pytest
